refactor: KatanaAppをドメイン別traitsに解体
- KatanaAppの巨大なimplブロック(shell.rs)を解体
- Action, Workspace, Export, Download, Update, Preview, Documentの7つのドメイン固有traitに機能を分離
- shell_ui.rsのupdate_uiへの依存を整理
- linter制約およびコンパイルエラーを全解決
