+++ crates/katana-ui/src/main.rs
@@ -9,21 +9,9 @@
     clippy::unwrap_used,
     clippy::panic,
     clippy::todo,
-    clippy::unimplemented
+    clippy::unimplemented,
+    clippy::cognitive_complexity
 )]
-#![warn(
-    clippy::expect_used,
-    clippy::indexing_slicing,
-    clippy::missing_errors_doc
-)]
-#![allow(missing_docs)]
-#![allow(clippy::too_many_lines)]
-#![allow(clippy::missing_errors_doc)]
-#![allow(clippy::indexing_slicing)]
-#![allow(clippy::cognitive_complexity)]
-#![allow(clippy::unwrap_used)]
-#![allow(clippy::expect_used)]
-#![allow(clippy::panic)]
 //! KatanA UI application entry point.
 
 #[cfg(not(test))]
